Add BDD tests for Axiell contributors and subjects
