+++ catalogue_graph/tests/adapters/transformers/axiell/bdd/test_contributors.py
@@ -1,0 +1,3 @@
+from pytest_bdd import scenarios
+
+scenarios("features/contributors.feature")
