+++ catalogue_graph/tests/adapters/transformers/axiell/bdd/test_subjects.py
@@ -1,0 +1,3 @@
+from pytest_bdd import scenarios
+
+scenarios("features/subjects.feature")
